fix: issues from ent update

+++ app/user/contest/internal/biz/submit.go
@@ -11,13 +11,13 @@ type Submit struct {
 	UserID      int64     `json:"user_id,omitempty"`
 	ProblemID   int64     `json:"problem_id,omitempty"`
 	Code        string    `json:"code,omitempty"`
-	Status      int32     `json:"state,omitempty"`
-	Point       int32     `json:"point,omitempty"`
+	Status      int16     `json:"state,omitempty"`
+	Point       int16     `json:"point,omitempty"`
 	CreateTime  time.Time `json:"create_time"`
 	TotalTime   int32     `json:"total_time,omitempty"`
 	MaxMemory   int32     `json:"max_memory,omitempty"`
 	Language    string    `json:"language,omitempty"`
-	CaseVersion int32     `json:"case_version,omitempty"`
+	CaseVersion int8      `json:"case_version,omitempty"`
 }
 
 type Pretest struct {
@@ -30,7 +30,7 @@ type Pretest struct {
 
 type SubmitRepo interface {
 	CreateSubmit(ctx context.Context, submit *Submit) (int64, error)
-	UpdateStatus(ctx context.Context, submitID int64, status int32) error
+	UpdateStatus(ctx context.Context, submitID int64, status int16) error
 	UpdateSubmit(ctx context.Context, submit *Submit) error
 	GetSubmission(ctx context.Context, submitID int64, userID int64) (*Submit, error)
 	CreatePretest(ctx context.Context, pretest *Pretest) error
@@ -53,7 +53,7 @@ func (uc *SubmitUsecase) CreateSubmit(ctx context.Context, submit *Submit) (int6
 	return submitID, nil
 }
 
-func (uc *SubmitUsecase) UpdateStatus(ctx context.Context, submitID int64, status int32) error {
+func (uc *SubmitUsecase) UpdateStatus(ctx context.Context, submitID int64, status int16) error {
 	return uc.repo.UpdateStatus(ctx, submitID, status)
 }
